UUID notes + bootstrap

+++ app/Http/Controllers/NoteController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Note;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class NoteController extends Controller
 {
@@ -21,6 +23,7 @@ class NoteController extends Controller
     {
         $note = new Note();
         $note->title = $request->title;
+        $note->uuid = Str::uuid();
         $note->body = $request->body;
         $note->user_id = auth()->user()->id;
         $note->save();
@@ -32,6 +35,12 @@ class NoteController extends Controller
     // function to delete a note
     public function destroy(Note $note)
     {
+
+        //@todo update with gates and policies
+        if ( $note->user_id != Auth::id()) {
+            abort(403);
+        }
+
         $note->delete();
         return redirect()->route('notes.index');
     }
@@ -39,18 +48,36 @@ class NoteController extends Controller
     // function to view a note
     public function show(Note $note)
     {
+
+        //@todo update with gates and policies
+        if ( $note->user_id != Auth::id()) {
+            abort(403);
+        }
+
         return view('single', compact('note'));
     }
 
     // function to edit a note
     public function edit(Note $note)
     {
+
+        //@todo update with gates and policies
+        if ( $note->user_id != Auth::id()) {
+            abort(403);
+        }
+
         return view('edit', compact('note'));
     }
 
     // function to update a note
     public function update(Request $request, Note $note)
     {
+
+        //@todo update with gates and policies
+        if ( $note->user_id != Auth::id()) {
+            abort(403);
+        }
+
         $note->title = $request->title;
         $note->body = $request->body;
         $note->save();
